+++ src/main/resources/application.properties
@@ -41,7 +41,7 @@ spring.datasource.hikari.connection-test-query=SELECT 1
 # Liquibase
 spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.xml
 spring.liquibase.enabled=false
-spring.liquibase.default-schema=dev
+spring.liquibase.default_schema=dev
 spring.liquibase.parameters.avenirs.admin.role=avenirs_api_admin_role
 security.authentication.filter=hmac
 # Jasypt
